convert download progress to compose dialog
