+++ app/src/main/kotlin/com/pnuema/bible/android/ui/versionselection/state/DownloadProgress.kt
@@ -1,9 +1,8 @@
-package com.pnuema.bible.android.ui.versionselection.dialogs
+package com.pnuema.bible.android.ui.versionselection.state
 
 data class DownloadProgress (
     val versionToDownload: String = "",
     val versionToDownloadDisplay: String = "",
     val max: Int = Int.MAX_VALUE,
     val progress: Int = 0,
-    val isComplete: Boolean = false,
 )
