+++ app/src/main/kotlin/com/pnuema/bible/android/ui/versionselection/ui/compose/VersionSelectionScreen.kt
@@ -1,22 +1,30 @@
 package com.pnuema.bible.android.ui.versionselection.ui.compose
 
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.LinearProgressIndicator
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Dialog
 import com.pnuema.bible.android.R
 import com.pnuema.bible.android.data.IVersion
 import com.pnuema.bible.android.data.firefly.VersionsDomain
 import com.pnuema.bible.android.database.VersionOffline
 import com.pnuema.bible.android.statics.CurrentSelected
 import com.pnuema.bible.android.ui.BibleTheme
+import com.pnuema.bible.android.ui.versionselection.state.DownloadProgress
 import com.pnuema.bible.android.ui.versionselection.state.VersionsState
 
 @Composable
@@ -40,6 +48,12 @@ fun VersionSelectionScreen(
                     onActionClicked = onActionClicked,
                 )
             }
+        }
+
+        state.downloadProgress?.let {
+            VersionDownloadDialog(
+                state = it
+            )
         }
 
         state.downloadDialogVersion?.let {
@@ -97,6 +111,44 @@ fun VersionSelectionScreen(
                         Text(text = stringResource(id = R.string.download_confirm_no))
                     }
                 }
+            )
+        }
+    }
+}
+
+@Composable
+fun VersionDownloadDialog(
+    modifier: Modifier = Modifier,
+    state: DownloadProgress,
+) {
+    Dialog(
+        onDismissRequest = {},
+    ) {
+        Column(
+            modifier = modifier
+                .padding(
+                    vertical = 8.dp,
+                    horizontal = 16.dp
+                ),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text(
+                modifier = Modifier.padding(bottom = 16.dp),
+                color = MaterialTheme.colorScheme.onBackground,
+                textAlign = TextAlign.Center,
+                text = stringResource(
+                    R.string.downloading_version_x,
+                    state.versionToDownloadDisplay
+                )
+            )
+            LinearProgressIndicator(
+                modifier = Modifier
+                    .align(alignment = Alignment.CenterHorizontally)
+                    .padding(bottom = 2.dp),
+                progress = state.progress.toFloat() / state.max.toFloat(),
+                strokeCap = StrokeCap.Round,
+                color = MaterialTheme.colorScheme.secondary,
+                trackColor = MaterialTheme.colorScheme.secondaryContainer
             )
         }
     }
